C program for file parsing a file and strings
This C file is used to extract the words that begin with a given character(s), those characters are defined at the top of the program file under the constant definitions as "EXTRACTION_CHAR"s. The final output produces the word count for those characters and the words themselves.
The output can be copied to a file by using the following command in your terminal: "./a.out > FILENAME.txt".